AuthCheck assigned to persist user session
The AuthCheck stateless widget is in charge of validating if a user has a current session. The StreamBuilder takes a stream and a builder value. The sream is using onAuthStateChange which recieves a notification everytime an auth event happens. The builder takes in the context and a snapshot which then checks if the session is not null, to then return the HomeScreen widget; persisting the user's session. Otherwise, it returns the LoginScreen widget.